Update: プロフィールの編集画面をtabから/account/setting/profileに移行
