+++ front/app/_containers/account-setting/container.tsx
@@ -1,0 +1,11 @@
+'use server'
+
+import AccountSettingTopPresentation from '@/app/_containers/account-setting/presentation'
+
+export default async function AccountSettingTopContainer() {
+	return (
+		<>
+			<AccountSettingTopPresentation></AccountSettingTopPresentation>
+		</>
+	)
+}
