+++ front/app/_containers/account-setting/presentation.tsx
@@ -1,0 +1,10 @@
+/**
+ * @package
+ */
+export default function AccountSettingTopPresentation() {
+	return (
+		<div className="flex w-full h-full justify-center items-center">
+			<p className="text-subtext">左のサイドメニューから設定項目を選択することができます。</p>
+		</div>
+	)
+}
