+++ front/app/_containers/profile/headerPresentation.tsx
@@ -1,5 +1,10 @@
+'use client'
 import { UserType } from '@/action/type/user'
 import DPHeader from '@/components/layout/header'
+import DPButton from '@/components/ui/button/button'
+import TextWithIcon from '@/components/ui/text/textWithIcon'
+import { Link } from 'react-aria-components'
+import { RiPencilLine } from 'react-icons/ri'
 
 type Props = {
 	user: UserType | undefined
@@ -10,5 +15,15 @@ type Props = {
  * @package
  */
 export default function HeaderPresentation({ user, isLogin }: Props) {
-	return <DPHeader user={user} isLogin={isLogin}></DPHeader>
+	return (
+		<DPHeader user={user} isLogin={isLogin}>
+			{user && (
+				<Link href={`/account/setting/profile`} className="outline-none">
+					<DPButton colormode="primary">
+						<TextWithIcon icon={<RiPencilLine />}>プロフィール編集</TextWithIcon>
+					</DPButton>
+				</Link>
+			)}
+		</DPHeader>
+	)
 }
