+++ front/app/account/setting/page.tsx
@@ -1,0 +1,9 @@
+import AccountSettingTopContainer from '@/app/_containers/account-setting/container'
+
+export default function AccountSettingPage() {
+	return (
+		<>
+			<AccountSettingTopContainer></AccountSettingTopContainer>
+		</>
+	)
+}
