+++ front/app/account/setting/profile/page.tsx
@@ -1,0 +1,11 @@
+import ProfileSettingContainer from '@/app/_containers/account-setting/profile/container'
+
+const ProfileSettingPage = async () => {
+	return (
+		<>
+			<ProfileSettingContainer></ProfileSettingContainer>
+		</>
+	)
+}
+
+export default ProfileSettingPage
